+++ go/cmd/entware-stats/main.go
@@ -14,10 +14,17 @@ func main() {
 		fmt.Println("<p class='error'>ENDPOINT not set</p>")
 		return
 	}
-	if ep != "stats" {
+	switch ep {
+	case "stats":
+		stats.Handle()
+	case "version":
+		stats.HandleVersion()
+	case "help":
+		stats.HandleHelp()
+	case "links_load":
+		stats.HandleLinksLoad()
+	default:
 		fmt.Println("Content-type: text/html; charset=utf-8\n")
 		fmt.Printf("<p class='error'>unknown endpoint: %s</p>", ep)
-		return
 	}
-	stats.Handle()
 }